# Optimize metal: input expansion