--- a/src/neural/backends/metal/mps/MetalNetworkBuilder.mm
+++ b/src/neural/backends/metal/mps/MetalNetworkBuilder.mm
@@ -56,12 +56,6 @@ std::string MetalNetworkBuilder::init(int gpu_id)
     return std::string([devices[gpu_id].name UTF8String]);
 }
 
-bool MetalNetworkBuilder::isMacOsVersionOrNewer(int major, int minor) const
-{
-    NSOperatingSystemVersion version = [[NSProcessInfo processInfo] operatingSystemVersion];
-    return (version.majorVersion > major) || (version.majorVersion == major && version.minorVersion >= minor);
-}
-
 void MetalNetworkBuilder::build(int kInputPlanes, MultiHeadWeights& weights, InputEmbedding embedding,
                                 bool attn_body, bool attn_policy, bool conv_policy, bool wdl, bool moves_left,
                                 Activations& activations, std::string& policy_head, std::string& value_head)
@@ -73,30 +67,16 @@ void MetalNetworkBuilder::build(int kInputPlanes, MultiHeadWeights& weights, Inp
     NSString * policyHead = [NSString stringWithUTF8String:policy_head.c_str()];
     NSString * valueHead = [NSString stringWithUTF8String:value_head.c_str()];
 
-    // 0. Input placeholder.
-    // @todo - placeholder can be made directly as NHWC to avoid transposes.
-    MPSGraphTensor * layer;
-    if (@available(macOS 13.0, *)) {
-        // New input placeholder of bitboard values.
-        layer = [graph inputPlaceholderWithInputChannels:kInputPlanes
-                                                  height:1
-                                                   width:1
-                                                   label:@"inputs"];
-        // Create a mask placeholder for the new input.
-        MPSGraphTensor * maskTensor = [graph maskPlaceholderWithInputChannels:kInputPlanes
-                                                                        label:@"inputs/mask"];
+    // 0. Input value and mask placeholders.
+    MPSGraphTensor * layer = [graph inputPlaceholderWithInputChannels:kInputPlanes
+                                                                label:@"inputs"];
 
-        layer = [graph expandInputTensorWithMask:maskTensor
-                                           input:layer
-                                           label:@"inputs/expand"];
-    }
-    else {
-        // Old input placeholder without mask, used for MacOS versions before 13.0.
-        layer = [graph inputPlaceholderWithInputChannels:kInputPlanes
-                                                  height:8
-                                                   width:8
-                                                   label:@"inputs"];
-    }
+    MPSGraphTensor * maskTensor = [graph maskPlaceholderWithInputChannels:kInputPlanes
+                                                                    label:@"inputs/mask"];
+
+    layer = [graph expandInputTensorWithMask:maskTensor
+                                       input:layer
+                                       label:@"inputs/expand"];
 
     const NSUInteger kernelSize = 3;
     const bool isPeDenseEmbedding = embedding == InputEmbedding::INPUT_EMBEDDING_PE_DENSE;
@@ -322,14 +302,6 @@ void MetalNetworkBuilder::build(int kInputPlanes, MultiHeadWeights& weights, Inp
     }
     else {
         [graph setResultTensors:@[policy, value]];
-    }
-}
-
-void MetalNetworkBuilder::forwardEvalLegacy(float * inputs, int batchSize, std::vector<float *> output_mems)
-{
-    @autoreleasepool {
-        Lc0NetworkGraph * graph = [Lc0NetworkGraph getGraphAt:[NSNumber numberWithInt:this->gpu_id]];
-        [graph runInferenceWithBatchSize:batchSize inputs:inputs masks:nil outputs:&output_mems[0]];
     }
 }