--- a/src/neural/backends/metal/mps/NetworkGraph.h
+++ b/src/neural/backends/metal/mps/NetworkGraph.h
@@ -50,6 +50,7 @@ static MPSImageFeatureChannelFormat fcFormat = MPSImageFeatureChannelFormatFloat
 
     // Input tensor and tensor data placeholders.
     MPSGraphTensor * _inputTensor;
+    MPSGraphTensor * _maskTensor;
 
     // Variables to track results of graph inference.
     NSArray<MPSGraphTensor *> * _resultTensors;
@@ -71,10 +72,19 @@ static MPSImageFeatureChannelFormat fcFormat = MPSImageFeatureChannelFormatFloat
 
 -(nonnull instancetype) initWithDevice:(id<MTLDevice> __nonnull)device;
 
+-(bool) isMacOsVersionOrNewer:(int)major minor:(int)minor;
+
 -(nonnull MPSGraphTensor *) inputPlaceholderWithInputChannels:(NSUInteger)channels
                                                        height:(NSUInteger)height
                                                         width:(NSUInteger)width
                                                         label:(NSString * __nullable)label;
+
+-(nonnull MPSGraphTensor *) maskPlaceholderWithInputChannels:(NSUInteger)channels
+                                                       label:(NSString * __nullable)label;
+
+-(nonnull MPSGraphTensor *) expandInputTensorWithMask:(MPSGraphTensor * __nonnull)maskTensor
+                                                input:(MPSGraphTensor * __nonnull)inputTensor
+                                                label:(NSString * __nonnull)label;
 
 -(nonnull MPSGraphTensor *) addConvolutionBlockWithParent:(MPSGraphTensor * __nonnull)parent
                                            outputChannels:(NSUInteger)outputChannels
@@ -199,9 +209,11 @@ static MPSImageFeatureChannelFormat fcFormat = MPSImageFeatureChannelFormatFloat
 
 -(nonnull NSArray<MPSGraphTensor *> *) runInferenceWithBatchSize:(NSUInteger)batchSize
                                                           inputs:(float * __nonnull)inputs
+                                                           masks:(uint64_t * __nullable)masks
                                                          outputs:(float * __nonnull * __nonnull)outputBuffers;
 
 -(nonnull MPSCommandBuffer *) runCommandSubBatchWithInputs:(float * __nonnull)inputs
+                                                     masks:(uint64_t * __nullable)masks
                                                   subBatch:(NSUInteger)subBatch
                                               subBatchSize:(NSUInteger)subBatchSize;