--- a/src/neural/backends/metal/mps/NetworkGraph.mm
+++ b/src/neural/backends/metal/mps/NetworkGraph.mm
@@ -1519,5 +1519,3 @@ static const NSInteger kMinSubBatchSize = 20;
 }
 
 @end
-
-