Recent additions to MPSGraph from macos 13.0 have made it possible to implement several performance improvements in the metal backend. This is the first in a series of improvements that would cover the following:

1. Move input bitboard expansion to the gpu
2. Implement policy mapping on the gpu (was not possible in 12.x due to a bug in GatherND implementation)
3. Compile the execution graph to speed up subsequent evaluations