@almaudoh 2025-07-25
```
Benchmark batch size 40 with inference average time 263.646ms - throughput 151.718 nps.
Benchmark batch size 60 with inference average time 385.377ms - throughput 155.692 nps.
Benchmark batch size 80 with inference average time 500.938ms - throughput 159.7 nps.
Benchmark batch size 100 with inference average time 630.121ms - throughput 158.7 nps.
Benchmark batch size 120 with inference average time 745.847ms - throughput 160.891 nps.
Benchmark batch size 140 with inference average time 873.982ms - throughput 160.186 nps.
```