@borg323 2025-07-25
> ```
> Benchmark batch size 40 with inference average time 263.646ms - throughput 151.718 nps.
> ```

Is this with the old or the new version?