@almaudoh 2025-07-25
Did another benchmark with `11248` so I can get higher batchsizes, where I expect a bigger difference.
**11248**
New version
```
Benchmark batch size 900 with inference average time 1152.38ms - throughput 780.989 nps.
Benchmark batch size 920 with inference average time 1168.41ms - throughput 787.393 nps.
Benchmark batch size 940 with inference average time 1186.42ms - throughput 792.3 nps.
Benchmark batch size 960 with inference average time 1215.92ms - throughput 789.523 nps.
Benchmark batch size 980 with inference average time 1240.48ms - throughput 790.018 nps.
Benchmark batch size 1000 with inference average time 1267.94ms - throughput 788.681 nps.
```

Old version
```
Benchmark batch size 900 with inference average time 1164.04ms - throughput 773.169 nps.
Benchmark batch size 920 with inference average time 1191.71ms - throughput 771.997 nps.
Benchmark batch size 940 with inference average time 1211.71ms - throughput 775.763 nps.
Benchmark batch size 960 with inference average time 1236.46ms - throughput 776.412 nps.
Benchmark batch size 980 with inference average time 1264.95ms - throughput 774.733 nps.
```