@almaudoh 2025-07-26
> I don't like having 'forwardEvalLegacy', but not that much.

I will remove it in a separate PR, once I've got all the other optimizations in, and have figured out a good way to remove it.